@@ -1,6 +1,6 @@
 module github.com/nwg-piotr/nwg-dock
 
-go 1.16
+go 1.19
 
 require (
 	github.com/allan-simon/go-singleinstance v0.0.0-20210120080615-d0997106ab37
@@ -8,4 +8,12 @@ require (
 	github.com/gotk3/gotk3 v0.6.1
 	github.com/joshuarubin/go-sway v0.0.4
 	github.com/sirupsen/logrus v1.8.1
+)
+
+require (
+	github.com/joshuarubin/lifecycle v1.0.0 // indirect
+	go.uber.org/atomic v1.3.2 // indirect
+	go.uber.org/multierr v1.1.0 // indirect
+	golang.org/x/sync v0.0.0-20190412183630-56d357773e84 // indirect
+	golang.org/x/sys v0.0.0-20191026070338-33540a1f6037 // indirect
 )
